requirements.txt
async-timeout==3.0.1
asyncio==3.4.3
chardet==3.0.4
DateTime==4.3
discord.py
numpy
Pillow==8.1.0
requests==2.25.1
wikipedia==1.4.0
psycopg2

